+++ stream-chat-android-client/src/main/java/io/getstream/chat/android/client/api/ChatApi.kt
@@ -276,6 +276,13 @@ internal interface ChatApi {
     ): Call<Unit>
 
     @CheckResult
+    fun markThreadRead(
+        channelType: String,
+        channelId: String,
+        threadId: String,
+    ): Call<Unit>
+
+    @CheckResult
     fun markUnread(
         channelType: String,
         channelId: String,