+++ CHANGELOG.md
@@ -22,6 +22,8 @@
 - Avoid multiple `ChatClient::connectUser` invocations by sharing the `Call`. [#5439](https://github.com/GetStream/stream-chat-android/pull/5439)
 
 ### ✅ Added
+- Add `ChatClient::markThreadRead` to mark a given thread as read. [#5447](https://github.com/GetStream/stream-chat-android/pull/5447)
+- Add `ChannelClient::markThreadRead` to mark a given thread as read. [#5447](https://github.com/GetStream/stream-chat-android/pull/5447)
 
 ### ⚠️ Changed
 
@@ -44,6 +46,7 @@
 ### ⬆️ Improved
 
 ### ✅ Added
+- Add `GlobalState::unreadThreadsCount` property to observe the number of unread threads. [#5452](https://github.com/GetStream/stream-chat-android/pull/5452)
 
 ### ⚠️ Changed